+++ src/components/image.py
@@ -10,7 +10,7 @@ class Image(Component):
             surface = pygame.image.fromstring(PIL_img.tobytes(), PIL_img.size, PIL_img.mode)
         else:
             if not basePath:
-                path = pathlib.Path(__file__).parent.parent.parent.absolute()
+                path = pathlib.Path(__file__).parent.parent.absolute()
                 path = path.joinpath("assets/")
                 path = str(path.joinpath(source))
